requirements.txt
openai
python-dotenv
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
